requirements.txt
discord.py
uvicorn
flask
pytube
PyNaCl
ffmpeg-python
google-api-python-client
google-auth
